@@ -3,25 +3,27 @@
   "version": "0.1.0",
   "scripts": {
     "ng": "ng",
-    "dev": "ng serve --proxy-config dev.proxy.conf.json",
+    "dev": "ng serve --proxy-config dev.proxy.conf.json ",
     "start": "ng serve --proxy-config proxy.conf.json",
-    "build": "ng build --configuration production",
+    "build": "npm run updateBuild -- $MDC_LIGHT_WEBSERVER_VERSION && ng build --configuration production",
+    "build-dev": "ng build --configuration development",
     "watch": "ng build --watch --configuration development",
-    "test": "ng test"
+    "test": "ng test",
+    "updateBuild": "node ./replace.build.js"
   },
   "private": true,
   "dependencies": {
-    "@angular-material-components/datetime-picker": "^6.0.3",
-    "@angular/animations": "~12.2.0",
-    "@angular/cdk": "^12.2.4",
-    "@angular/common": "~12.2.0",
-    "@angular/compiler": "~12.2.0",
-    "@angular/core": "~12.2.0",
-    "@angular/forms": "~12.2.0",
-    "@angular/material": "^12.2.4",
-    "@angular/platform-browser": "~12.2.0",
-    "@angular/platform-browser-dynamic": "~12.2.0",
-    "@angular/router": "~12.2.0",
+    "@angular-material-components/datetime-picker": "~8.0.0",
+    "@angular/animations": "~14.0.4",
+    "@angular/cdk": "^14.0.4",
+    "@angular/common": "~14.0.4",
+    "@angular/compiler": "~14.0.4",
+    "@angular/core": "~14.0.4",
+    "@angular/forms": "~14.0.4",
+    "@angular/material": "^14.0.4",
+    "@angular/platform-browser": "~14.0.4",
+    "@angular/platform-browser-dynamic": "~14.0.4",
+    "@angular/router": "~14.0.4",
     "@juggle/resize-observer": "^3.3.1",
     "@ngx-translate/core": "^13.0.0",
     "@ngx-translate/http-loader": "^6.0.0",
@@ -33,20 +35,19 @@
     "moment-timezone": "^0.5.34",
     "ng": "^0.0.0",
     "ngx-echarts": "^7.0.2",
-    "ngx-ip": "^1.1.4",
-    "ngx-mat-select-search": "^3.3.2",
+    "ngx-mat-select-search": "^4.2.1",
     "ngx-toastr": "^14.1.3",
-    "ngx-webstorage": "^8.0.0",
+    "ngx-webstorage": "^10.0.1",
     "rxjs": "~6.6.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.2.4",
-    "@angular/cli": "~12.2.4",
-    "@angular/compiler-cli": "~12.2.0",
+    "@angular-devkit/build-angular": "^14.1.0",
+    "@angular/cli": "~14.0.4",
+    "@angular/compiler-cli": "~14.0.4",
     "@types/jasmine": "~3.8.0",
-    "@types/node": "^12.11.1",
+    "@types/node": "^18.0.6",
     "jasmine-core": "~3.8.0",
     "karma": "~6.3.0",
     "karma-chrome-launcher": "~3.1.0",
@@ -54,6 +55,14 @@
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
     "ng-swagger-gen": "^2.3.1",
-    "typescript": "~4.3.5"
+    "replace-in-file": "^6.3.5",
+    "typescript": "~4.7.4"
+  },
+  "engines": {
+    "npm": ">=8.3.0"
+  },
+  "overrides": {
+    "@angular-material-components/datetime-picker": "~8.0.0",
+    "highlight.js": "^10.0.0"
   }
 }
