@@ -28,5 +28,21 @@ export default [
   { name: 'sequencedAlarmAlarmNo', address: '/Nck/SequencedAlarms/alarmNo' },
   { name: 'alarmEventAlarmNo', address: '/Nck/AlarmEvent/alarmNo' },
   { name: 'lastAlarmNo', address: '/Nck/LastAlarm/alarmNo' },
-  { name: 'topPrioAlarmNo', address: '/Nck/TopPrioAlarm/alarmNo' }
+  { name: 'topPrioAlarmNo', address: '/Nck/TopPrioAlarm/alarmNo' },
+  { name: 'cuttingTime', address: '/Channel/ChannelDiagnose/cuttingTime'},
+  { name: 'progStatus', address: '/Channel/State/progStatus'},
+  { name: 'msg', address: '/Channel/ProgramInfo/msg'},
+  { name: 'workPNameLong', address: '/Channel/ProgramInfo/workPNameLong'},
+  { name: 'workPandProgName', address: '/Channel/ProgramInfo/workPandProgName'},
+  { name: 'specParts ', address: '/Channel/State/specParts'},
+  { name: 'turnState', address: '/Channel/Spindle/turnState'},
+  { name: 'anLanguageOnHmi', address: '/Nck/Configuration/anLanguageOnHmi'},
+  { name: 'basicLengthUnit', address: '/Nck/Configuration/basicLengthUnit'},
+  { name: 'nckType', address: '/Nck/Configuration/nckType'},
+  { name: 'nckVersion', address: '/Nck/Configuration/nckVersion'},
+  { name: 'chanStatus ', address: '/Channel/State/chanStatus'},
+  { name: 'optStopActive', address: '/Channel/ProgramModification/optStopActive'},
+  { name: 'progTestActive', address: '/Channel/ProgramModification/progTestActive'},
+  { name: 'trialRunActive', address: '/Channel/ProgramModification/trialRunActive'},
+  
 ];
