@@ -1,4 +1,5 @@
 export const environment = {
   production: true,
-  apiRoot: '/api/v1'
+  apiRoot: '/api/v1',
+  version: '{BUILD_VERSION}' // changes here must not be committed!
 };
