@@ -37,8 +37,19 @@ async function networkConfigGetHandler(
     return [undefined, undefined];
   });
 
-  const { x1: cx1, x2: cx2, time } = configManager.config.networkConfig;
+  const {
+    x1: cx1,
+    x2: cx2,
+    time: cfgTime
+  } = configManager.config.networkConfig;
 
+  const reachable = await TimeManager.testNTPServer(cfgTime.ntpHost)
+    .then(() => true)
+    .catch(() => false);
+  const time = {
+    ...cfgTime,
+    reachable
+  };
   const merged = {
     x1: {
       useDhcp: x1?.dhcp || cx1.useDhcp,
@@ -87,14 +98,18 @@ async function networkConfigPatchHandler(
       // ISO8601 to YYYY-MM-DD hh:mm:ss
       const [YYYY, MM, DD, hh, mm, ss] =
         timeConfig.currentTime.split(/[/:\-T]/);
-      timePromise = TimeManager.setTimeManually(
+      timePromise = TimeManager.setTimeAndZoneManually(
         `${YYYY}-${MM}-${DD} ${hh}:${mm}:${ss.slice(0, 2)}`,
         timeConfig.timezone
       );
     } else {
-      const allReadyConfigured = (timeConfig.ntpHost === configManager.config.networkConfig.time?.ntpHost );
-      if (allReadyConfigured) winston.debug(`${logPrefix} received ntp-config again, ignore data.`);
-      timePromise =  allReadyConfigured ? Promise.resolve() : TimeManager.setNTPServer(timeConfig.ntpHost);
+      const allReadyConfigured =
+        timeConfig.ntpHost === configManager.config.networkConfig.time?.ntpHost;
+      if (allReadyConfigured)
+        winston.debug(`${logPrefix} received ntp-config again, ignore data.`);
+      timePromise = allReadyConfigured
+        ? Promise.resolve()
+        : TimeManager.setNTPServer(timeConfig.ntpHost);
     }
   }
 
@@ -103,24 +118,25 @@ async function networkConfigPatchHandler(
     NetworkManagerCliController.setConfiguration('eth0', x1Config),
     NetworkManagerCliController.setConfiguration('eth1', x2Config),
     timePromise
-  ]).then((results) => {
-    results.forEach((result) => {
-      if (result.status === 'rejected') {
-        if (result.reason && (typeof result.reason === 'string') && result.reason.includes('is not available or is not a valid NTP server.')) {
-          errorMsg = result.reason;
-          return;
+  ])
+    .then((results) => {
+      results.forEach((result) => {
+        if (result.status === 'rejected') {
+          winston.error(
+            `networkConfigPatchHandler error due to ${result.reason}. Only writing configuration to config file.`
+          );
         }
-        winston.error(
-          `networkConfigPatchHandler error due to ${result.reason}. Only writing configuration to config file.`
-        );
-      }
+      });
+    })
+    .then(() => {
+      configManager.saveConfig({ networkConfig: request.body });
+      response
+        .status(errorMsg ? 400 : 200)
+        .json(errorMsg || configManager.config.networkConfig);
+    })
+    .catch((err) => {
+      winston.error(`${logPrefix} error due to ${err?.msg}`);
     });
-  }).then(() => {
-    configManager.saveConfig({ networkConfig: request.body });
-    response.status(errorMsg ? 400 : 200).json(errorMsg ?? configManager.config.networkConfig);
-  }).catch((err) => {
-    winston.error(`${logPrefix} error due to ${err?.msg}`);
-  }); 
 }
 
 export const networkConfigHandlers = {
