@@ -1,8 +1,10 @@
 import { ConfigManager } from '../../../../../ConfigManager';
 import { Request, response, Response } from 'express';
 import { v4 as uuidv4 } from 'uuid';
+import { VirtualDataPointManager } from '../../../../../VirtualDataPointManager';
 
 let configManager: ConfigManager;
+let vdpManager: VirtualDataPointManager;
 
 /**
  * Set ConfigManager to make accessible for local function
@@ -10,6 +12,10 @@ let configManager: ConfigManager;
  */
 export function setConfigManager(config: ConfigManager) {
   configManager = config;
+}
+
+export function setVdpManager(config: VirtualDataPointManager) {
+  vdpManager = config;
 }
 
 /**
@@ -34,6 +40,15 @@ async function vdpsPostHandler(
 ): Promise<void> {
   //TODO: Input validation
   const newData = { ...request.body, ...{ id: uuidv4() } };
+  if (
+    newData.operationType === 'counter' &&
+    newData.resetSchedules?.length > 0
+  ) {
+    for (const [index, resetEntry] of newData.resetSchedules.entries()) {
+      newData.resetSchedules[index].created = Date.now();
+      newData.resetSchedules[index].lastReset = undefined;
+    }
+  }
   configManager.changeConfig('insert', 'virtualDataPoints', newData);
   await configManager.configChangeCompleted();
   response.status(200).json({
@@ -53,6 +68,7 @@ async function vdpsPostBulkHandler(
 ): Promise<void> {
   try {
     await configManager.bulkChangeVirtualDataPoints(request.body || {});
+    await configManager.configChangeCompleted();
 
     response.status(200).send();
   } catch {
@@ -94,25 +110,29 @@ async function vdpDeleteHandler(
  * @param  {Request} request
  * @param  {Response} response
  */
- async function vdpPatchHandler(
+async function vdpPatchHandler(
   request: Request,
   response: Response
 ): Promise<void> {
-  configManager.changeConfig(
-    'update',
-    'virtualDataPoints',
-    request.body,
-    (vdp) => {
-      return (vdp.id = request.body.id);
-    }
-  );
-  await configManager.configChangeCompleted();
+  if (request.body.reset && request.params.id) {
+    vdpManager.resetCounter(request.params.id);
+    delete request.body.reset;
+  } else {
+    configManager.changeConfig(
+      'update',
+      'virtualDataPoints',
+      request.body,
+      (vdp) => {
+        return (vdp.id = request.body.id);
+      }
+    );
+    await configManager.configChangeCompleted();
+  }
   response.status(200).json({
     changed: request.body,
     href: `/vdps/${request.body.id}`
   });
 }
-
 
 export const virtualDatapointHandlers = {
   vdpsGet: vdpsGetHandler,
