@@ -1,6 +1,11 @@
-import { IErrorEvent, ILifecycleEvent } from '../../common/interfaces';
+import {
+  DataSourceProtocols,
+  IErrorEvent,
+  ILifecycleEvent
+} from '../../common/interfaces';
 import { EventBus } from '../EventBus';
 import { OPCUAServerOptions } from 'node-opcua';
+import { ScheduleDescription } from '../CounterManager';
 
 export interface IAuthConfig {
   secret: string;
@@ -11,6 +16,22 @@ export interface IAuthRuntimeConfig {
   expiresIn: number;
 }
 
+type IRegistry = {
+  [component in TSoftwareComponents]: {
+    tag: string;
+  };
+} & {
+  url: string;
+};
+
+type TSoftwareComponents = 'web' | 'mdc' | 'mtc';
+
+interface IRegistries {
+  prod: IRegistry;
+  dev: IRegistry;
+  stag: IRegistry;
+}
+
 export interface IRuntimeConfig {
   users: IUser[];
   mtconnect: IMTConnectConfig;
@@ -19,6 +40,7 @@ export interface IRuntimeConfig {
   auth: IAuthRuntimeConfig;
   datahub: IDataHubConfig;
   systemInfo: ISystemInfo[];
+  registries: IRegistries;
 }
 
 export interface IGeneralConfig {
@@ -78,7 +100,7 @@ export interface IDataPointConfig {
 export interface IDataSourceConfig {
   name: string;
   dataPoints: IDataPointConfig[];
-  protocol: string;
+  protocol: DataSourceProtocols;
   connection?: IS7DataSourceConnection;
   enabled: boolean;
   type?: IS7DataSourceTypes | IIoShieldDataSourcesTypes;
@@ -240,12 +262,35 @@ export interface QuickStartConfig {
 export function isDataPointMapping(obj: any): obj is IDataPointMapping {
   return 'source' in obj && 'target' in obj && !('id' in obj);
 }
-
+export interface EnumOperationEntry {
+  priority: number;
+  source: string;
+  returnValueIfTrue: string;
+}
 export interface IVirtualDataPointConfig {
   id: string;
   sources: string[];
-  operationType: 'and' | 'or' | 'not' | 'counter' | 'thresholds';
+  operationType:
+    | 'and'
+    | 'or'
+    | 'not'
+    | 'counter'
+    | 'thresholds'
+    | 'enumeration'
+    | 'greater'
+    | 'greaterEqual'
+    | 'smaller'
+    | 'smallerEqual'
+    | 'equal'
+    | 'unequal'
+    | 'sum';
   thresholds?: ITargetDataMap;
+  enumeration?: {
+    defaultValue?: string;
+    items: EnumOperationEntry[];
+  };
+  comparativeValue?: string | number;
+  resetSchedules?: ScheduleDescription[];
 }
 
 export interface ISystemInfoItem {
@@ -265,6 +310,14 @@ export interface TermsAndConditionsConfig {
   accepted: boolean;
 }
 
+type env = {
+  [component in TSoftwareComponents]: {
+    tag: string;
+  };
+} & {
+  selected: 'prod' | 'dev' | 'stag';
+};
+
 export interface IConfig {
   dataSources: IDataSourceConfig[];
   dataSinks: Array<IDataSinkConfig>;
@@ -274,6 +327,7 @@ export interface IConfig {
   networkConfig: NetworkConfig;
   quickStart: QuickStartConfig;
   termsAndConditions: TermsAndConditionsConfig;
+  env: env;
 }
 
 export interface IConfigManagerParams {
