@@ -6,10 +6,13 @@ import SshService from '../../SshService';
  */
 export class TimeManager {
   static CONFIG_PATH = '/etc/systemd/timesyncd.conf';
+
   /**
    * Parse network address out of config file.
    */
   private static parseConfig(configContent: string) {
+    const logPrefix = `TimeManager::parseConfig`;
+    winston.debug(`${logPrefix} called with ${configContent}`);
     const lines = configContent.split('\n');
     for (const line of lines) {
       if (line.indexOf('NTP=') >= 0) {
@@ -22,27 +25,34 @@ export class TimeManager {
       }
     }
   }
+
   /**
    * Compose new ntp server to config string.
    */
   private static composeConfig(
     configContent: string,
-    ntpServerIP: string
+    ntpServerAddress: string
   ): string {
+    const logPrefix = `TimerManager::composeConfig`;
+    winston.debug(
+      `${logPrefix} called with new ntp server: ${ntpServerAddress}`
+    );
     let newConfig = '';
     for (const line of configContent.split('\n')) {
       if (line.indexOf('NTP=') >= 0 && line.indexOf('FallbackNTP=') < 0) {
-        if (ntpServerIP === '0.0.0.0') {
+        if (ntpServerAddress === '0.0.0.0') {
           newConfig += `NTP=\n`;
         } else {
-          newConfig += `NTP=${ntpServerIP}\n`;
+          newConfig += `NTP=${ntpServerAddress}\n`;
         }
       } else {
         newConfig += `${line}\n`;
       }
     }
+    winston.debug(`${logPrefix} new composed config: \n${newConfig}`);
     return newConfig;
   }
+
   /**
    * Restart time service on host.
    */
@@ -50,7 +60,7 @@ export class TimeManager {
     const logPrefix = `TimeManager::restartTimesyncdService`;
     winston.debug(`${logPrefix} restarting...`);
     const command = `systemctl restart systemd-timesyncd`;
-    const { stdout, stderr } = await SshService.sendCommand(command);
+    const { stdout, stderr } = await SshService.sendCommand(command, true);
     if (stderr !== '') {
       const msg = `${logPrefix} error during restart of timesyncd service.`;
       winston.error(msg);
@@ -58,19 +68,21 @@ export class TimeManager {
     }
     winston.debug(`${logPrefix} restart successfully.`);
   }
+
   /**
    * Set NTP Server on remote host.
    */
-  public static async setNTPServer(ntpServerIP: string): Promise<void> {
+  public static async setNTPServer(ntpServerAddress: string): Promise<void> {
     const logPrefix = `TimeManager::setNTPServer`;
+    winston.debug(`${logPrefix} `);
     try {
-      await this.testNTPServer(ntpServerIP)
+      await this.testNTPServer(ntpServerAddress);
     } catch (error) {
       return Promise.reject(
-        `${ntpServerIP} is not available or is not a valid NTP server.`
+        `${ntpServerAddress} is not available or is not a valid NTP server.`
       );
     }
-    
+
     const readCommand = `cat ${TimeManager.CONFIG_PATH}`;
 
     let { stdout: currentConfig, stderr } = await SshService.sendCommand(
@@ -81,61 +93,134 @@ export class TimeManager {
       winston.error(msg);
       return Promise.reject(new Error(msg));
     }
+    if (typeof currentConfig !== 'string') {
+      winston.error(`${logPrefix} expect string but receive buffer. Abort.`);
+      return Promise.reject();
+    }
     winston.info(
       `${logPrefix} read current config from host: ${currentConfig}`
     );
+    // Config file is empty -> Bugfix reset to default
     if (currentConfig.trim() === '') {
-      currentConfig = `#  This file is part of systemd.
-#
-#  systemd is free software; you can redistribute it and/or modify it
-#  under the terms of the GNU Lesser General Public License as published by
-#  the Free Software Foundation; either version 2.1 of the License, or
-#  (at your option) any later version.
-#
-# Entries in this file show the compile time defaults.
-# You can change settings by editing this file.
-# Defaults can be restored by simply deleting this file.
-#
-# See timesyncd.conf(5) for details.
-
-[Time]
-#NTP=
-#FallbackNTP=0.debian.pool.ntp.org 1.debian.pool.ntp.org 2.debian.pool.ntp.org 3.debian.pool.ntp.org
-#RootDistanceMaxSec=5
-#PollIntervalMinSec=32
-#PollIntervalMaxSec=2048`;
+      currentConfig = [
+        '# This file is part of systemd.',
+        '#',
+        '#  systemd is free software; you can redistribute it and/or modify it',
+        '#  under the terms of the GNU Lesser General Public License as published by',
+        '#  the Free Software Foundation; either version 2.1 of the License, or',
+        '#  (at your option) any later version.',
+        '#',
+        '# Entries in this file show the compile time defaults.',
+        '# You can change settings by editing this file.',
+        '# Defaults can be restored by simply deleting this file.',
+        '#',
+        '# See timesyncd.conf(5) for details.',
+        '',
+        '[Time]',
+        '#NTP=',
+        '#FallbackNTP=0.debian.pool.ntp.org 1.debian.pool.ntp.org 2.debian.pool.ntp.org 3.debian.pool.ntp.org',
+        '#RootDistanceMaxSec=5',
+        '#PollIntervalMinSec=32',
+        '#PollIntervalMaxSec=2048'
+      ].join('\n');
     }
-    const newConfig = await this.composeConfig(currentConfig, ntpServerIP);
+    const newConfig = await this.composeConfig(currentConfig, ntpServerAddress);
     winston.info(`${logPrefix} generate new config: ${newConfig}`);
-    const writeCommand = `echo "${newConfig}" > ${this.CONFIG_PATH}`;
-    const t = await SshService.sendCommand(writeCommand);
-    winston.info(
-      `${logPrefix} written new NTP-Server (${ntpServerIP}) to host`
-    );
-    await this.restartTimesyncdService();
+    const writeCommand = `tee ${this.CONFIG_PATH} <<EOF
+${newConfig.trim()}\nEOF`;
+    return SshService.sendCommand(writeCommand)
+      .then((response) => {
+        if (response.stderr.length > 0) {
+          winston.error(
+            `${logPrefix} error during write to ${this.CONFIG_PATH} due to ${response.stderr}`
+          );
+          return Promise.reject(response.stderr);
+        }
+        winston.info(
+          `${logPrefix} written new NTP-Server (${ntpServerAddress}) to host.`
+        );
+        return this.restartTimesyncdService();
+      })
+      .catch((err) => {
+        winston.error(
+          `${logPrefix} error setting ntp server due to ${JSON.stringify(err)}`
+        );
+      });
   }
+
   /**
    * Return currently configured NTP server.
    */
   public static async getNTPServer(): Promise<string> {
     const logPrefix = `TimeManager::getNTPServer`;
-    winston.info(`${logPrefix} reading config from remote host.`);
+    winston.info(`${logPrefix} reading config from host machine.`);
     const readCommand = `cat ${this.CONFIG_PATH}`;
-    const currentConfig = await SshService.sendCommand(readCommand);
-    winston.info(
-      `${logPrefix} read config from remote host successful. ${currentConfig}`
+    return await SshService.sendCommand(readCommand).then((response) => {
+      if (response.stderr.length > 0) {
+        winston.error(
+          `${logPrefix} error reading ntp server due to ${response.stderr}`
+        );
+        return Promise.reject(response.stderr);
+      }
+      if (typeof response.stdout !== 'string') {
+        winston.error(`${logPrefix} expect string but got buffer. Abort`);
+        return Promise.reject();
+      }
+      winston.info(
+        `${logPrefix} read config from remote host successful. ${response.stdout}`
+      );
+      return this.parseConfig(response.stdout);
+    });
+  }
+
+  /**
+   * Setting time via timedatectl
+   * @param time
+   * @returns
+   */
+  public static async setTimeManually(time: string) {
+    const logPrefix = `TimeManager::setTimeManually`;
+
+    await SshService.sendCommand(`timedatectl set-ntp 0`, true);
+    winston.info(`${logPrefix} disabling NTP successfully.`);
+    winston.info(`${logPrefix} setting time to ${time}`);
+    const setTimeSshCommand = `timedatectl set-time '${time}'`;
+    const setTimeRes = await SshService.sendCommand(setTimeSshCommand, true);
+    if (setTimeRes.stderr !== '') {
+      const errMsg = `${logPrefix} error during set manual time due to ${setTimeRes.stderr}`;
+      winston.error(errMsg);
+      return Promise.reject(errMsg);
+    }
+  }
+
+  /**
+   * Setting time zone via timedatectl
+   * @param timezone
+   * @returns
+   */
+  public static async setTimeZoneManually(timezone: string) {
+    const logPrefix = `TimeManager::setTimeZoneManually`;
+    winston.info(`${logPrefix} setting time zone to ${timezone}`);
+    const setTimezoneSshCommand = `timedatectl set-timezone '${timezone}'`;
+    const setTimezoneRes = await SshService.sendCommand(
+      setTimezoneSshCommand,
+      true
     );
-    return this.parseConfig(currentConfig);
+    if (setTimezoneRes.stderr !== '') {
+      const errMsg = `${logPrefix} error during set manual time due to ${setTimezoneRes.stderr}`;
+      winston.error(errMsg);
+      return Promise.reject(errMsg);
+    }
   }
 
   /**
    * Set time manually and disable NTP.
    */
-  public static async setTimeManually(
+  public static async setTimeAndZoneManually(
     time: string,
     timezone: string
   ): Promise<void> {
-    const logPrefix = `TimeManager::setTimeManually`;
+    const logPrefix = `TimeManager::setTimeAndZoneManually`;
     winston.info(`${logPrefix} ${time} in timezone ${timezone}`);
     const timeRegex =
       /^\d{4}\-(0[1-9]|1[012])\-(0[1-9]|[12][0-9]|3[01]) (0[1-9]|1[0-9]|2[01234]):([0-5]{1}[0-9]{1}|60):([0-5]{1}[0-9]{1}|60)/;
@@ -146,31 +231,18 @@ export class TimeManager {
     }
     // Disable NTP
     winston.info(`${logPrefix} disabling NTP.`);
-    await SshService.sendCommand(`timedatectl set-ntp 0`);
-    winston.info(`${logPrefix} disabling NTP successfully.`);
-    winston.info(`${logPrefix} setting time to ${time}`);
-    const setTimeSshCommand = `timedatectl set-time "${time}"`;
-    const setTimeRes = await SshService.sendCommand(setTimeSshCommand);
-    if (setTimeRes.stderr !== '') {
-      const errMsg = `${logPrefix} error during set manual time due to ${setTimeRes.stderr}`;
-      winston.error(errMsg);
-      return Promise.reject(errMsg);
-    }
-    winston.info(`${logPrefix} setting time zone to ${time}`);
-    const setTimezoneSshCommand = `timedatectl set-timezone "${timezone}"`;
-    const setTimezoneRes = await SshService.sendCommand(setTimezoneSshCommand);
-    if (setTimezoneRes.stderr !== '') {
-      const errMsg = `${logPrefix} error during set manual time due to ${setTimeRes.stderr}`;
-      winston.error(errMsg);
-      return Promise.reject(errMsg);
-    }
+
+    // Time Zone must be set first to prevent wrong set time!!!
+    await TimeManager.setTimeZoneManually(timezone);
+    await TimeManager.setTimeManually(time);
+
     winston.info(`${logPrefix} set time ${time} ${timezone} successfully.`);
   }
 
   /**
    * Test if a server is a valid and available ntp server.
    */
-  private static async testNTPServer(server: string): Promise<void> {
+  public static async testNTPServer(server: string): Promise<void> {
     const logPrefix = `TimeManager::testNTPServer`;
     winston.debug(`${logPrefix} start testing: ${server}`);
 
