@@ -3,12 +3,28 @@ import { ConfigManager } from '../ConfigManager';
 import { IVirtualDataPointConfig } from '../ConfigManager/interfaces';
 import { CounterManager } from '../CounterManager';
 import { DataPointCache } from '../DatapointCache';
-import { IDataSourceMeasurementEvent } from '../Southbound/DataSources/interfaces';
+import {
+  IDataSourceMeasurementEvent,
+  IMeasurement
+} from '../Southbound/DataSources/interfaces';
+import { SynchronousIntervalScheduler } from '../SyncScheduler';
 
 interface IVirtualDataPointManagerParams {
   configManager: ConfigManager;
   cache: DataPointCache;
 }
+
+type LogEntry = {
+  count: number;
+  log: string;
+};
+
+type LogSummary = {
+  error: LogEntry[];
+  warn: LogEntry[];
+  info: LogEntry[];
+  debug: LogEntry[];
+};
 
 /**
  * Calculates virtual datapoints
@@ -18,21 +34,83 @@ export class VirtualDataPointManager {
   private config: IVirtualDataPointConfig[] = null;
   private cache: DataPointCache;
   private counters: CounterManager;
+  private scheduler: SynchronousIntervalScheduler;
+  private logSchedulerListenerId: number;
+  private logSummary: LogSummary = {
+    error: [],
+    warn: [],
+    info: [],
+    debug: []
+  };
 
   private static className: string = VirtualDataPointManager.name;
 
   constructor(params: IVirtualDataPointManagerParams) {
-    params.configManager.once('configsLoaded', () => {
-      return this.init();
-    });
+    params.configManager.once('configsLoaded', () => this.init());
+    params.configManager.on('configChange', () => this.updateConfig());
 
     this.configManager = params.configManager;
     this.cache = params.cache;
-    this.counters = new CounterManager();
+    this.counters = new CounterManager(this.configManager);
+    this.scheduler = SynchronousIntervalScheduler.getInstance();
   }
 
   private init() {
+    this.updateConfig();
+    this.setupLogCycle();
+  }
+
+  private updateConfig() {
+    const logPrefix = `${VirtualDataPointManager.className}::updateConfig`;
+    winston.debug(`${logPrefix} refreshing config`);
+
     this.config = this.configManager.config.virtualDataPoints;
+  }
+
+  /**
+   * Setup log cycle for summary logging
+   */
+  protected setupLogCycle(defaultFrequency: number = 15 * 60 * 1000): void {
+    if (this.logSchedulerListenerId) return;
+
+    const logPrefix = `${VirtualDataPointManager.className}::setupLogCycle`;
+    winston.debug(`${logPrefix} setup log cycle`);
+
+    this.logSchedulerListenerId = this.scheduler.addListener(
+      [defaultFrequency],
+      this.printSummaryLogs.bind(this)
+    );
+  }
+
+  /**
+   * Logs summary of logs to not spam logs every data point ready cycle
+   */
+  protected printSummaryLogs(): void {
+    const logPrefix = `${VirtualDataPointManager.className}::logSummary`;
+
+    Object.keys(this.logSummary).forEach((level: keyof LogSummary) => {
+      if (this.logSummary[level].length > 0) {
+        for (const entry of this.logSummary[level]) {
+          winston[level](
+            `${logPrefix} Count: ${entry.count} Log: ${entry.log}`
+          );
+        }
+        this.logSummary[level] = [];
+      }
+    });
+  }
+
+  /**
+   * Adding log to summary. If Same log already exits, counting up
+   */
+  protected addSummaryLog(level: keyof LogSummary, log: string): void {
+    const entry = this.logSummary[level].find((entry) => entry.log === log);
+
+    if (entry) {
+      entry.count = entry.count + 1;
+    } else {
+      this.logSummary[level].push({ log, count: 1 });
+    }
   }
 
   /**
@@ -51,6 +129,45 @@ export class VirtualDataPointManager {
   }
 
   /**
+   * Calculates the sum of a  list of virtual data points.
+   *
+   * @param  {IDataSourceMeasurementEvent[]} sourceEvents
+   * @param  {IVirtualDataPointConfig} config
+   * @returns boolean
+   */
+  private sum(
+    sourceEvents: IDataSourceMeasurementEvent[],
+    config: IVirtualDataPointConfig
+  ): number | null {
+    const logPrefix = `VirtualDataPointManager::sum`;
+    if (config.operationType !== 'sum') {
+      winston.error(`${logPrefix} called with invalid operation type.`);
+      return null;
+    }
+    if (sourceEvents.length < 2) {
+      winston.warn(
+        `${logPrefix} virtual data point (${config.id}) requires at least 2 sources!`
+      );
+      return null;
+    }
+    try {
+      return sourceEvents.reduce(
+        (previousValue, { measurement: { value } }) => {
+          if (typeof value !== 'number') {
+            // Not addable
+            throw Error(`${value} is not a number.`);
+          }
+          return value + previousValue;
+        },
+        0
+      );
+    } catch (err) {
+      winston.error(`${logPrefix} ${err.message}`);
+      return null;
+    }
+  }
+
+  /**
    * Calculates virtual data points from type and
    *
    * @param  {IDataSourceMeasurementEvent[]} sourceEvents
@@ -62,7 +179,8 @@ export class VirtualDataPointManager {
     config: IVirtualDataPointConfig
   ): boolean | null {
     if (sourceEvents.length < 2) {
-      winston.warn(
+      this.addSummaryLog(
+        'warn',
         `Virtual data point (${config.id}) requires at least 2 sources!`
       );
       return null;
@@ -77,6 +195,229 @@ export class VirtualDataPointManager {
   }
 
   /**
+   * Check if value of sourceEvents first entry is greater the compare value config.comparativeValue.
+   *
+   * @param sourceEvents  Array of measurements
+   * @param config operation config
+   * @param equal optional parameter to check for >= instead of >
+   * @returns true if it is greater, false if not and null if something went wrong.
+   */
+  private greater(
+    sourceEvents: IDataSourceMeasurementEvent[],
+    config: IVirtualDataPointConfig,
+    equal: boolean = false
+  ): boolean | null {
+    const logPrefix = `${this.constructor.name}::greater${
+      equal ? '(Equal)' : ''
+    }`;
+    if (sourceEvents.length > 1 || sourceEvents.length === 0) {
+      this.addSummaryLog(
+        'warn',
+        `${logPrefix} is only available for one source but receive ${sourceEvents.length}`
+      );
+      return null;
+    }
+
+    // @ts-ignore
+    const compare = parseFloat(config.comparativeValue);
+    if (Number.isNaN(compare)) {
+      this.addSummaryLog(
+        'warn',
+        `${logPrefix} got ${compare} for comparison but it's not a number`
+      );
+      return null;
+    }
+
+    const value = sourceEvents[0].measurement.value;
+
+    let result: boolean;
+    switch (typeof value) {
+      case 'number': {
+        if (equal && Math.abs(compare - value) < 1e-9) {
+          return true;
+        }
+        return value > compare;
+      }
+      case 'string': {
+        const parsed = parseFloat(value);
+        if (isNaN(parsed)) {
+          this.addSummaryLog('error', `${logPrefix} no valid number.`);
+          return null;
+        }
+        if (equal && Math.abs(compare - parsed) < 1e-9) {
+          return true;
+        }
+        return parsed > compare;
+      }
+      case 'boolean': {
+        this.addSummaryLog(
+          'warn',
+          `${logPrefix} boolean is not a valid compare value.`
+        );
+        return null;
+      }
+      default: {
+        this.addSummaryLog(
+          'warn',
+          `${logPrefix} invalid value type: ${typeof value}`
+        );
+        return null;
+      }
+    }
+  }
+
+  /**
+   * Check if value of sourceEvents first entry is smaller the compare value config.comparativeValue.
+   *
+   * @param sourceEvents  Array of measurements
+   * @param config operation config
+   * @param equal optional parameter to check for <= instead of <
+   * @returns true if it is smaller, false if not and null if something went wrong.
+   */
+  private smaller(
+    sourceEvents: IDataSourceMeasurementEvent[],
+    config: IVirtualDataPointConfig,
+    equal: boolean = false
+  ): boolean | null {
+    const logPrefix = `${this.constructor.name}::smaller ${
+      equal ? '(Equal)' : ''
+    }`;
+    if (sourceEvents.length > 1 || sourceEvents.length === 0) {
+      this.addSummaryLog(
+        'warn',
+        `${logPrefix} is only available for one source but receive ${sourceEvents.length}`
+      );
+      return null;
+    }
+
+    // @ts-ignore
+    const compare = parseFloat(config.comparativeValue);
+    if (Number.isNaN(compare)) {
+      this.addSummaryLog(
+        'warn',
+        `${logPrefix} got ${compare} for comparison but it's not a number`
+      );
+      return null;
+    }
+
+    const value = sourceEvents[0].measurement.value;
+    let result: boolean;
+    switch (typeof value) {
+      case 'number': {
+        // Correct method to compare
+        // floating-point numbers
+        if (equal && Math.abs(compare - value) < 1e-9) {
+          return true;
+        }
+        return value < compare;
+      }
+      case 'string': {
+        const parsed = parseFloat(value);
+        if (isNaN(parsed)) {
+          this.addSummaryLog('error', `${logPrefix} no valid number.`);
+          return null;
+        }
+        if (equal && Math.abs(compare - parsed) < 1e-9) {
+          return true;
+        }
+        return parsed < compare;
+      }
+      case 'boolean': {
+        this.addSummaryLog(
+          'warn',
+          `${logPrefix} boolean is not a valid compare value.`
+        );
+        return null;
+      }
+      default: {
+        this.addSummaryLog(
+          'warn',
+          `${logPrefix} invalid value type: ${typeof value}`
+        );
+        return null;
+      }
+    }
+  }
+
+  /**
+   * Check if value of sourceEvents first entry is equal the compare value config.comparativeValue.
+   *
+   * @param sourceEvents  Array of measurements
+   * @param config operation config
+   * @returns true if it is equal, false if not and null if something went wrong.
+   */
+  private equal(
+    sourceEvents: IDataSourceMeasurementEvent[],
+    config: IVirtualDataPointConfig
+  ): boolean | null {
+    const logPrefix = `${this.constructor.name}::equal`;
+    if (sourceEvents.length > 1 || sourceEvents.length === 0) {
+      this.addSummaryLog(
+        'warn',
+        `${logPrefix} is only available for one source but receive ${sourceEvents.length}`
+      );
+      return null;
+    }
+    const value = sourceEvents[0].measurement.value;
+    const compare = config.comparativeValue;
+    switch (typeof value) {
+      case 'number': {
+        // Correct method to compare
+        // floating-point numbers
+
+        if (
+          // @ts-ignore
+          !Number.isNaN(parseFloat(compare)) &&
+          // @ts-ignore
+          Math.abs(parseFloat(compare) - value) < 1e-9
+        ) {
+          return true;
+        } else {
+          return false;
+        }
+      }
+      case 'string': {
+        if (typeof compare !== 'string') {
+          this.addSummaryLog(
+            'warn',
+            `${logPrefix} try to compare string with non string value: ${compare}`
+          );
+          return null;
+        }
+        return value.trim() === compare.trim();
+      }
+      case 'boolean': {
+        this.addSummaryLog(
+          'warn',
+          `${logPrefix} boolean is not a valid compare value.`
+        );
+        return null;
+      }
+      default: {
+        this.addSummaryLog(
+          'warn',
+          `${logPrefix} invalid value type: ${typeof value}`
+        );
+        return null;
+      }
+    }
+  }
+
+  /**
+   * Check if value of sourceEvents first entry is unequal the compare value config.comparativeValue.
+   *
+   * @param sourceEvents  Array of measurements
+   * @param config operation config
+   * @returns true if it is unequal, false if not and null if something went wrong.
+   */
+  private unequal(
+    sourceEvents: IDataSourceMeasurementEvent[],
+    config: IVirtualDataPointConfig
+  ): boolean | null {
+    return !this.equal(sourceEvents, config);
+  }
+
+  /**
    * Calculates virtual data points from type or
    *
    * @param  {IDataSourceMeasurementEvent[]} sourceEvents
@@ -88,7 +429,8 @@ export class VirtualDataPointManager {
     config: IVirtualDataPointConfig
   ): boolean | null {
     if (sourceEvents.length < 2) {
-      winston.warn(
+      this.addSummaryLog(
+        'warn',
         `Virtual data point (${config.id}) requires at least 2 sources!`
       );
       return null;
@@ -114,7 +456,8 @@ export class VirtualDataPointManager {
     config: IVirtualDataPointConfig
   ): boolean | null {
     if (sourceEvents.length !== 1) {
-      winston.warn(
+      this.addSummaryLog(
+        'warn',
         `Virtual data point (${config.id}) requires exactly 1 source!`
       );
       return null;
@@ -135,7 +478,8 @@ export class VirtualDataPointManager {
     config: IVirtualDataPointConfig
   ): number | null {
     if (sourceEvents.length !== 1) {
-      winston.warn(
+      this.addSummaryLog(
+        'warn',
         `Virtual data point (${config.id}) requires exactly 1 source!`
       );
       return null;
@@ -146,10 +490,60 @@ export class VirtualDataPointManager {
       this.toBoolean(measurement.value) &&
       this.cache.hasChanged(measurement.id)
     ) {
-      return this.counters.increment(measurement.id);
+      return this.counters.increment(config.id);
     }
 
     return null;
+  }
+
+  /**
+   * Check 'enumerated' grouped virtual data point and return the 'returnValueIfTrue' of the highest true value.
+   *
+   * @param sourceEvents
+   * @param config
+   * @returns string
+   */
+  private enumeration(
+    sourceEvents: IDataSourceMeasurementEvent[],
+    config: IVirtualDataPointConfig
+  ): string | null {
+    const logPrefix = `${this.constructor.name}::enumeration`;
+    if (config.operationType !== 'enumeration') {
+      this.addSummaryLog(
+        'error',
+        `${logPrefix} receive invalid operation type: ${config.operationType}`
+      );
+      return null;
+    }
+    if (typeof config.enumeration === undefined) {
+      this.addSummaryLog(
+        'error',
+        `${logPrefix} no enumeration configuration found`
+      );
+      return null;
+    }
+
+    // Iterate over sorted by high prio array
+    for (const entry of config.enumeration.items.sort((a, b) => {
+      return b.priority - a.priority;
+    })) {
+      const hit = !!sourceEvents.find((event) => {
+        // winston.debug(
+        //   `${logPrefix} searching for entry ${entry.source} found ${event.measurement.id}`
+        // );
+        // winston.debug(
+        //   `${logPrefix} measurement is: ${!!event.measurement.value}`
+        // );
+        return (
+          event.measurement.id === entry.source && !!event.measurement.value
+        );
+      });
+      if (hit) {
+        return entry.returnValueIfTrue;
+      }
+    }
+    // No true value in list
+    return config.enumeration?.defaultValue || null;
   }
 
   /**
@@ -164,7 +558,8 @@ export class VirtualDataPointManager {
     config: IVirtualDataPointConfig
   ): number | null {
     if (sourceEvents.length !== 1) {
-      winston.warn(
+      this.addSummaryLog(
+        'warn',
         `Virtual data point (${config.id}) requires exactly 1 source!`
       );
       return null;
@@ -194,10 +589,10 @@ export class VirtualDataPointManager {
    * @param  {IVirtualDataPointConfig} config
    * @returns boolean
    */
-  private calulateValue(
+  private calculateValue(
     sourceEvents: IDataSourceMeasurementEvent[],
     config: IVirtualDataPointConfig
-  ): boolean | number | null {
+  ): IMeasurement['value'] {
     switch (config.operationType) {
       case 'and':
         return this.and(sourceEvents, config);
@@ -209,10 +604,27 @@ export class VirtualDataPointManager {
         return this.count(sourceEvents, config);
       case 'thresholds':
         return this.thresholds(sourceEvents, config);
+      case 'enumeration':
+        return this.enumeration(sourceEvents, config);
+      case 'greater':
+        return this.greater(sourceEvents, config);
+      case 'greaterEqual':
+        return this.greater(sourceEvents, config, true);
+      case 'smaller':
+        return this.smaller(sourceEvents, config);
+      case 'smallerEqual':
+        return this.smaller(sourceEvents, config, true);
+      case 'equal':
+        return this.equal(sourceEvents, config);
+      case 'unequal':
+        return this.unequal(sourceEvents, config);
+      case 'sum':
+        return this.sum(sourceEvents, config);
       default:
-        winston.warn(
-          `Invalid type (${config.operationType}) provided for virtual data point ${config.id}!`
-        );
+        // TODO Only print this once at startup or config change, if invalid
+        // this.addSummaryLog("warn",
+        //   `Invalid type (${config.operationType}) provided for virtual data point ${config.id}!`
+        // );
         return null;
     }
   }
@@ -229,7 +641,8 @@ export class VirtualDataPointManager {
   ): IDataSourceMeasurementEvent[] {
     const logPrefix = `${VirtualDataPointManager.className}::getVirtualEvents`;
     if (this.config === null) {
-      winston.warn(
+      this.addSummaryLog(
+        'warn',
         `${logPrefix} Config not yet loaded. Skipping virtual event calculation`
       );
       return [];
@@ -255,7 +668,8 @@ export class VirtualDataPointManager {
           event = this.cache.getCurrentEvent(sourceId);
         }
         if (!event) {
-          winston.warn(
+          this.addSummaryLog(
+            'warn',
             `Virtual data point ${vdpConfig.id} could not be calculated, because missing event from data source ${sourceId}`
           );
         }
@@ -266,7 +680,7 @@ export class VirtualDataPointManager {
       // Skip virtual data point if one or more source events are missing
       if (sourceEvents.some((event) => typeof event === 'undefined')) continue;
 
-      const value = this.calulateValue(sourceEvents, vdpConfig);
+      const value = this.calculateValue(sourceEvents, vdpConfig);
 
       if (value === null) {
         continue;
@@ -291,5 +705,14 @@ export class VirtualDataPointManager {
     this.cache.update(virtualEvents);
 
     return virtualEvents;
+  }
+
+  /**
+   * Set counter to 0
+   * @param id identifier of the virtual datapoint
+   */
+  public resetCounter(id: string): void {
+    winston.info(`${this.constructor.name}::resetCounter `);
+    this.counters.reset(id);
   }
 }
